requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
boto3==1.20.26
botocore==1.23.26
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cssselect2==0.4.1
dj-database-url==0.5.0
Django==4.0
django-address==0.2.5
django-bootstrap4==21.1
django-brotli==0.2.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-filter==21.1
django-formtools==2.3
django-heroku==0.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-storages==1.12.3
django-tables2==2.4.1
django-weasyprint==2.1.0
docopt==0.6.2
fonttools==4.28.5
gunicorn==20.1.0
html5lib==1.1
idna==3.3
jmespath==0.10.0
Pillow==8.4.0
pipreqs==0.4.11
psycopg2-binary==2.9.2
pycparser==2.21
pydyf==0.1.2
pyphen==0.11.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tinycss2==1.1.1
urllib3==1.26.7
weasyprint==53.4
webencodings==0.5.1
whitenoise==5.3.0
yarg==0.1.9
